add animation

+++ TodoList/Views/ListView.swift
@@ -10,16 +10,23 @@ import SwiftUI
 struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     var body: some View {
-        List{
-            ForEach(listViewModel.items, content: {
-                item in ListRowView(item: item).onTapGesture {
-                    withAnimation(.linear){
-                        listViewModel.updateItem(item: item)
-                    }
-                }
-            }).onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-        }.listStyle(PlainListStyle()).navigationTitle("Todo List")
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoItemView().transition(AnyTransition.opacity.animation(.easeInOut))
+            }else{
+                List{
+                    ForEach(listViewModel.items, content: {
+                        item in ListRowView(item: item).onTapGesture {
+                            withAnimation(.linear){
+                                listViewModel.updateItem(item: item)
+                            }
+                        }
+                    }).onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }.listStyle(PlainListStyle())
+            }
+        }
+        .navigationTitle("Todo List")
         .navigationBarItems(leading: EditButton(), trailing: NavigationLink("Add", destination: AddView()))
     }
